XMonad.Layout.OnHost allows host-specific modifications to a layout, which is otherwise very difficult to do.  Similarly to X.L.PerWorkspace, it provides onHost, onHosts, modHost, and modHosts layout modifiers.  It attempts to do smart hostname comparison, such that short names will be matched with short names and FQDNs with FQDNs.
This module currently requires that $HOST be set in the environment.
You can use System.Posix.Env.setEnv to do so in xmonad.hs if need be.
(Properly, this should be done via the network library, but I'm trying to
avoid adding that dependency.)  An alternative would be to shell out to
get the name, but that has considerable portability hurdles.